fixed unsigned long problem between 64 bit and 32 bit

+++ trunk/N6LYT/ghpsdr3/trunk/src/monitor/iqthread.h
@@ -27,7 +27,7 @@
 #define IQPORT 13000
 
 typedef struct _buffer {
-    unsigned long sequence;
+    unsigned long long sequence;
     unsigned short offset;
     unsigned short length;
     unsigned char data[500];
