+++ trunk/N6LYT/ghpsdr3/trunk/src/receiver/ozy.h
@@ -69,7 +69,7 @@ extern "C" {
 #define LT2208_RANDOM_ON          0x10
 
 typedef struct _buffer {
-    unsigned long sequence;
+    unsigned long long sequence;
     unsigned short offset;
     unsigned short length;
     unsigned char data[500];
